Extract SAN UPN as an attribute or fallback principal
